Continuing Work on Tab 4 - plot and tabular display of model evaluations.
ui.R: Added radio buttons giving choice of drawing plots as points,
lines, or both to be consistent with appearance of other plots.

serverTab2.R:  Cosmetic formatting changes plus a correction that
allows the data set name to appear in the plot title.

serverTab4: Made a correction that allows the data set name to appear
in the plot title.

PlotModelEvals.R:  This file was adapted from PlotModelResults.R.
Removed code that runs models to generate predictions, and changed
titles of plots.  Changed “data view” drop down menu from model results
such as MVF and IF to times to types of model evaluations such as
ln(prequential likelihood) and model bias.